Fix nullability mismatch of TResult of ValueTask<TResult> which would sometimes  cause a build error.
Sometimes a "rebuild all" would work with no error, sometimes not. This appears to fix the issue. The error suggested that a nullability match was required for method resolution (or extension method in this case), which I did not expect. The error seen was:

LAMA0611: Error CS1061 in code generated by aspect 'Cache': 'ValueTask<IAsyncEnumerable<int>?>' does not
contain a definition for 'AsAsyncEnumerable' and no accessible extension method 'AsAsyncEnumerable' accepting a first argument of type 'ValueTask<IAsyncEnumerable<int>?>'